+++ workspace/solutions/rust/src/main.rs
@@ -13,7 +13,4 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-pub mod example;
-pub mod json;
-pub mod structure;
-pub mod test;
+fn main() {}