+++ workspace/solutions/rust/src/s0/s0/s0/s1/mod.rs
@@ -1,2 +1,4 @@
 pub mod s_0001_00;
 pub mod s_0001_test;
+
+pub type TSolution = fn(Vec<i32>, i32) -> Vec<i32>;