+++ workspace/solutions/rust/src/s0/s0/s0/s1/s_0001_00.rs
@@ -23,3 +23,5 @@ impl Solution {
     }
 }
 // @lc code=end
+
+pub const SOLUTION: super::TSolution = Solution::two_sum;